Manage dependencies just with shadow-cljs
If we use deps.edn, we need to install `clojure` in the build environment. This could be simple, but
before I attempt to do that I want to see if just leveraging `shadow-cljs.edn`, and thus skipping
the call to `clojure`, will suffice.

We'll see if the build environment has a version of the JDK installed.